Enhance Hermes integration with NeMo-Relay and update documentation
- Added lifecycle note in README.md regarding manual file pruning in long-lived sessions.
- Updated Dockerfile comments to clarify the purpose of sitecustomize.py and its role in monkey-patching Slack token handling.
- Improved finalize-hook script with a TODO for upstream retirement when NeMo-Relay supports per-turn session finalization.
- Removed obsolete nemoclaw_patches.py file, streamlining the integration with Slack command handling.
- Refactored sitecustomize.py to simplify Slack command routing and improve error handling.
